[Refactor] extension 따로 빼기, 함수명 동사형으로 변경
